bePy/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
scikit-learn==1.4.0
openai==1.8.0
joblib==1.3.2
pandas==2.1.4
numpy==1.26.2
aiohttp==3.9.1